Reorganize the packages
