--- plantseg/main/__init__.py
@@ -1,5 +1,0 @@
-import logging
-
-gui_logger = logging.getLogger("PlantSeg")
-# hardcode the log-level for now
-gui_logger.setLevel(logging.INFO)
