+++ plantseg/pipeline/__init__.py
@@ -1,0 +1,5 @@
+import logging
+
+gui_logger = logging.getLogger("PlantSeg")
+# hardcode the log-level for now
+gui_logger.setLevel(logging.INFO)
